// src/app/interfaces/enums.ts
export enum positions {
  topRight = "top-right",
  topCenter = "top-center",
  topLeft = "top-left",
  topFullWidth = "top-full-width",
  bottomRight = "bottom-right",
  bottomCenter = "bottom-center",
  bottomLeft = "bottom-left",
  bottomFullWidth = "bottom-full-width",
}

export enum animates {
  slideFromBottom = "slideFromBottom",
  slideFromTop = "slideFromTop",
  slideFromRight = "slideFromRight",
  slideFromLeft = "slideFromLeft",
  fade = "fade",
}